Reorganize files and logic, as well as add build/run scripts
