+++ README.md
@@ -1,6 +1,7 @@
 TODO: Everything.
 
 [![Build status](https://ci.appveyor.com/api/projects/status/9jfet757ah3w7mrr/branch/develop?svg=true)](https://ci.appveyor.com/project/VariantXYZ/polariscore)
+
 [![Build status](https://travis-ci.org/PolarisTeam/PolarisCore.svg?branch=develop)](https://travis-ci.org/PolarisTeam/PolarisCore)
 
 Going to move and re-organize all the code from the PolarisServer project and get it to work with .NET core for 
@@ -9,3 +10,7 @@ multi-platform shenanigans.
 develop should be where all the untested (but compileable) code, file PRs to this
 
 master should be where all tested/stable/no warning code goes, don't commit here directly, merge changes from develop
+
+To Build:
+
+Run build.sh inside the PolarisCore root directory
