+++ build.sh
@@ -1,0 +1,1 @@
+dotnet publish -o ./bin ./src/PolarisServer
