+++ run.sh
@@ -1,0 +1,1 @@
+dotnet bin/PolarisServer.dll
