+++ src/Polaris.sln
@@ -3,11 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
 VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "PolarisServer", "PolarisServer\PolarisServer.xproj", "{59B03D6A-617A-4A81-93FD-0CDAB5030169}"
-EndProject
-Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "PolarisAuth", "PolarisAuth\PolarisAuth.xproj", "{085E9274-34D6-4C7D-9E97-92DB76FF040A}"
-EndProject
 Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "PolarisLib", "PolarisLib\PolarisLib.xproj", "{EE343089-6170-43B4-B2CA-65B1223EBC15}"
+EndProject
+Project("{8BB2217D-0F2D-49D1-97BC-3654ED321F3B}") = "PolarisServer", "PolarisServer\PolarisServer.xproj", "{085E9274-34D6-4C7D-9E97-92DB76FF040A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,18 +13,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{59B03D6A-617A-4A81-93FD-0CDAB5030169}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{59B03D6A-617A-4A81-93FD-0CDAB5030169}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{59B03D6A-617A-4A81-93FD-0CDAB5030169}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{59B03D6A-617A-4A81-93FD-0CDAB5030169}.Release|Any CPU.Build.0 = Release|Any CPU
-		{085E9274-34D6-4C7D-9E97-92DB76FF040A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{085E9274-34D6-4C7D-9E97-92DB76FF040A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{085E9274-34D6-4C7D-9E97-92DB76FF040A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{085E9274-34D6-4C7D-9E97-92DB76FF040A}.Release|Any CPU.Build.0 = Release|Any CPU
 		{EE343089-6170-43B4-B2CA-65B1223EBC15}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{EE343089-6170-43B4-B2CA-65B1223EBC15}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{EE343089-6170-43B4-B2CA-65B1223EBC15}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{EE343089-6170-43B4-B2CA-65B1223EBC15}.Release|Any CPU.Build.0 = Release|Any CPU
+		{085E9274-34D6-4C7D-9E97-92DB76FF040A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{085E9274-34D6-4C7D-9E97-92DB76FF040A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{085E9274-34D6-4C7D-9E97-92DB76FF040A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{085E9274-34D6-4C7D-9E97-92DB76FF040A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
