--- src/PolarisAuth/project.json
@@ -1,23 +1,0 @@
-{
-  "version": "1.0.0-*",
-  "buildOptions": {
-    "emitEntryPoint": true
-  },
-
-  "dependencies": {
-    "Microsoft.NETCore.App": {
-      "type": "platform",
-      "version": "1.0.1"
-    },
-    "Newtonsoft.Json": "9.0.1",
-    "PolarisLib": "*",
-    "System.Runtime.Loader": "4.0.0",
-    "System.Security.Cryptography.Csp": "4.0.0"
-  },
-
-  "frameworks": {
-    "netcoreapp1.0": {
-      "imports": "dnxcore50"
-    }
-  }
-}
