+++ src/PolarisLib/PolarisLib.xproj
@@ -10,7 +10,7 @@
     <ProjectGuid>ee343089-6170-43b4-b2ca-65b1223ebc15</ProjectGuid>
     <RootNamespace>PolarisLib</RootNamespace>
     <BaseIntermediateOutputPath Condition="'$(BaseIntermediateOutputPath)'=='' ">.\obj</BaseIntermediateOutputPath>
-    <OutputPath Condition="'$(OutputPath)'=='' ">.\bin\</OutputPath>
+    <OutputPath Condition="'$(OutputPath)'=='' ">..\..\bin\</OutputPath>
     <TargetFrameworkVersion>v4.5.2</TargetFrameworkVersion>
   </PropertyGroup>
