+++ src/global.json
@@ -1,6 +1,5 @@
 {
     "projects": [
-        "PolarisAuth",
         "PolarisLib",
         "PolarisServer"
     ]
